+++ apps/E2E/src/Input/consts.ts
@@ -1,0 +1,13 @@
+export const HOMEPAGE_INPUT_BUTTON = 'Homepage_Input_Button';
+export const INPUT_TESTPAGE = 'Input_TestPage';
+
+/* E2E Testing Input 1 */
+export const INPUT_TEST_COMPONENT = 'Input_Test_Component'; // A component on each specific test page
+export const INPUT_TEST_COMPONENT_DISMISS_BUTTON = 'Input_Test_Component_Dismiss_Button'; // A component on each specific test page
+
+export const INPUT_TEXT = 'Input_Text'; // For testing the text input functionality of input
+
+export const INPUT_START_STRING = 'Test';
+export const INPUT_TYPE_STRING = ' '; // For testing the typing functionality of input
+export const INPUT_ERROR_STRING = 'Input_Error_String'; // For testing the error functionality of input
+export const INPUT_ONCLICK_STRING = 'Input_Onclick_String'; // For testing the accessory icon button functionality of input