+++ packages/components/Menu/src/MenuCallout/index.ts
@@ -1,0 +1,3 @@
+export { MenuCallout } from './MenuCallout';
+export { menuCalloutName } from './MenuCallout.types';
+export type { MenuCalloutProps, MenuCalloutTokens } from './MenuCallout.types';