+++ packages/components/Menu/src/MenuList/MenuListTokens.android.ts
@@ -1,0 +1,11 @@
+import type { Theme } from '@fluentui-react-native/framework';
+import { globalTokens } from '@fluentui-react-native/theme-tokens';
+import type { TokenSettings } from '@fluentui-react-native/use-styling';
+
+import type { MenuListTokens } from './MenuList.types';
+
+export const defaultMenuListTokens: TokenSettings<MenuListTokens, Theme> = (t: Theme): MenuListTokens => ({
+  backgroundColor: t.colors.neutralBackground1,
+  cornerRadius: globalTokens.corner.radius80,
+  paddingVertical: globalTokens.size80,
+});