+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,116 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Mon, 10 Apr 2023 19:48:24 GMT and should not be manually modified.
+This log was last generated on Mon, 01 May 2023 20:43:14 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.143.1
+
+Mon, 01 May 2023 20:43:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/spinner to v0.4.0
+
+## 0.143.0
+
+Thu, 27 Apr 2023 19:45:00 GMT
+
+### Minor changes
+
+- set n of m on submenu trigger (krsiler@microsoft.com)
+- Bump @fluentui-react-native/menu to v1.8.0
+
+## 0.142.5
+
+Wed, 26 Apr 2023 20:11:44 GMT
+
+### Patches
+
+- fixed accessibility props in pressable test for press to alert button (gulnazsayed@microsoft.com)
+
+## 0.142.4
+
+Tue, 25 Apr 2023 02:09:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu to v1.7.0
+
+## 0.142.3
+
+Sat, 22 Apr 2023 23:41:05 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.32.51
+- Bump @fluentui-react-native/dropdown to v0.7.48
+- Bump @fluentui-react-native/experimental-button to v0.16.109
+- Bump @fluentui-react-native/experimental-menu-button to v0.6.50
+- Bump @fluentui-react-native/menu to v1.6.3
+- Bump @fluentui-react-native/notification to v0.21.53
+- Bump @fluentui/react-native to v0.36.23
+
+## 0.142.2
+
+Fri, 21 Apr 2023 04:54:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu to v1.6.2
+
+## 0.142.1
+
+Fri, 21 Apr 2023 00:04:59 GMT
+
+### Patches
+
+- enable pointer capture to be overridden (krsiler@microsoft.com)
+- Bump @fluentui-react-native/menu to v1.6.1
+
+## 0.142.0
+
+Thu, 20 Apr 2023 09:41:22 GMT
+
+### Minor changes
+
+- Expand input functionality (ayushsinghs@yahoo.in)
+- Bump @fluentui-react-native/input to v0.2.0
+
+## 0.141.1
+
+Wed, 19 Apr 2023 10:08:39 GMT
+
+### Patches
+
+- add screenshots and doc update (rohanpd.work@gmail.com)
+- Bump @fluentui-react-native/menu to v1.6.0
+
+## 0.141.0
+
+Wed, 12 Apr 2023 18:42:22 GMT
+
+### Minor changes
+
+- fix tester and radio complete (rohanpd.work@gmail.com)
+- Bump @fluentui-react-native/button to v0.32.50
+- Bump @fluentui-react-native/dropdown to v0.7.47
+- Bump @fluentui-react-native/experimental-button to v0.16.108
+- Bump @fluentui-react-native/experimental-checkbox to v0.13.51
+- Bump @fluentui-react-native/experimental-menu-button to v0.6.49
+- Bump @fluentui-react-native/input to v0.1.2
+- Bump @fluentui-react-native/menu to v1.5.0
+- Bump @fluentui-react-native/notification to v0.21.52
+- Bump @fluentui/react-native to v0.36.22
+
+## 0.140.0
+
+Tue, 11 Apr 2023 10:42:50 GMT
+
+### Minor changes
+
+- Add new component - input (ayushsinghs@yahoo.in)
+- Bump @fluentui-react-native/input to v0.1.1
 
 ## 0.139.0