+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Mon, 10 Apr 2023 19:48:24 GMT and should not be manually modified.
+This log was last generated on Sat, 22 Apr 2023 23:41:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.50
+
+Sat, 22 Apr 2023 23:41:05 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-button to v0.16.109
+
+## 0.6.49
+
+Wed, 12 Apr 2023 18:42:22 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-button to v0.16.108
 
 ## 0.6.48